Refactor database configuration to use environment variables and remove deprecated variables file
