+++ alembic.ini
@@ -61,7 +61,7 @@ version_path_separator = os  # Use os.pathsep. Default configuration used for ne
 # are written from script.py.mako
 # output_encoding = utf-8
 
-sqlalchemy.url = driver://user:pass@localhost/dbname
+sqlalchemy.url = postgresql://{USER_DB}:{PASSWORD_DB}@{HOST_DB}/{NAME_DB}
 
 
 [post_write_hooks]
