+++ alembic/versions/4b97b56966f8_primera_revision.py
@@ -1,0 +1,26 @@
+"""Primera revision
+
+Revision ID: 4b97b56966f8
+Revises: 
+Create Date: 2024-12-19 10:34:02.559574
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = '4b97b56966f8'
+down_revision: Union[str, None] = None
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    pass
+
+
+def downgrade() -> None:
+    pass
